Update test suites and add remote suite
