--- .github/workflows/build.yml
@@ -1,66 +1,0 @@
-name: Build with integration tests and code coverage
-
-on:
-  push:
-    branches: [ '17.0' ]
-  pull_request:
-    branches: [ '17.0' ]
-
-permissions:
-  checks: write
-  pull-requests: write
-  repository-projects: write
-  contents: write
-
-jobs:
-  build:
-    #    strategy:
-    #      matrix:
-    #        environment: [ AWS-1, YANDEX-1, YANDEX-2 ]
-    #        java-version: [ 17 ]
-    #    environment: ${{ matrix.environment }}
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6
-
-      - name: Set up JDK
-        uses: actions/setup-java@be666c2fcd27ec809703dec50e508c2fdc7f6654 # v5
-        with:
-          java-version: 17
-          distribution: 'liberica'
-          cache: 'gradle'
-
-      - name: Build with Gradle
-        run: ./gradlew check integrationTest jacocoTestReport
-
-      - name: Publish Test Report
-        uses: dorny/test-reporter@a43b3a5f7366b97d083190328d2c652e1a8b6aa2 # v3.0.0
-        if: ${{ !cancelled() }}
-        with:
-          name: Test Results
-          path: '**/build/test-results/**/*.xml'
-          reporter: java-junit
-          token: ${{ secrets.GITHUB_TOKEN }}
-
-      - name: Jacoco Report
-        id: jacoco
-        uses: madrapps/jacoco-report@50d3aff4548aa991e6753342d9ba291084e63848 # v1.7.2
-        if: ${{ !cancelled() }}
-        with:
-          paths: >-
-            ${{ github.workspace }}/modules/commons-vfs/build/reports/jacoco/test/jacocoTestReport.xml,
-            ${{ github.workspace }}/modules/jdk/build/reports/jacoco/test/jacocoTestReport.xml,
-            ${{ github.workspace }}/modules/spring-6/build/reports/jacoco/test/jacocoTestReport.xml
-          token: ${{ secrets.GITHUB_TOKEN }}
-          min-coverage-overall: 40
-          min-coverage-changed-files: 60
-          title: Code Coverage
-          comment-type: ${{ github.event_name == 'pull_request' && 'pr_comment' || 'summary' }}
-          update-comment: true
-
-#      - name: Fail PR if overall coverage is less than 80%
-#        if: ${{ steps.jacoco.outputs.coverage-overall < 80.0 }}
-#        uses: actions/github-script@v8
-#        with:
-#          script: |
-#            core.setFailed('Overall coverage is less than 80%!')
