+++ modules/commons-vfs/src/integrationTest/java/com/github/vfss3/commonsvfs/S3IntegrationContext.java
@@ -9,10 +9,11 @@ import org.apache.commons.vfs2.FileSystemOptions;
 
 /**
  * Process-wide data holder shared between a backend-specific {@code @Suite} (LocalStack,
- * MinIO, …) and the test cases in {@link com.github.vfss3.commonsvfs.tests}.
+ * MinIO, … or a remote AWS / Yandex bucket) and the test cases in
+ * {@link com.github.vfss3.commonsvfs.tests}.
  *
- * <p>The suite calls {@link #initialize(URI, S3FileSystemOptions)} from {@code @BeforeSuite}
- * and {@link #reset()} from {@code @AfterSuite}. Test classes read {@link #rootUrl()} and
+ * <p>The suite calls one of the {@code initialize} overloads from {@code @BeforeSuite} and
+ * {@link #reset()} from {@code @AfterSuite}. Test classes read {@link #rootUrl()} and
  * {@link #options()} and resolve VFS themselves — this class deliberately has no
  * dependency on Apache Commons VFS.
  */
@@ -32,7 +33,8 @@ public final class S3IntegrationContext {
 
     /**
      * Build a unique bucket URL for this run on the supplied endpoint and store the
-     * (URL, options) pair for the test cases to read.
+     * (URL, options) pair for the test cases to read. Used by the local container suites,
+     * which can freely create and delete throwaway buckets.
      */
     public static void initialize(URI endpoint, S3FileSystemOptions backendOptions) {
         Objects.requireNonNull(endpoint, "endpoint");
@@ -46,7 +48,19 @@ public final class S3IntegrationContext {
         String bucketName = "vfs3-tests-" + token;
 
         // Custom-endpoint (path-style) URL: s3://host:port/bucket/
-        rootUrl = String.format("s3://%s:%d/%s/", endpoint.getHost(), endpoint.getPort(), bucketName);
+        initialize(String.format("s3://%s:%d/%s/", endpoint.getHost(), endpoint.getPort(), bucketName), backendOptions);
+    }
+
+    /**
+     * Use the supplied bucket URL verbatim (no random bucket generation, no CreateBucket
+     * call). Intended for remote suites that target a pre-existing AWS / Yandex / etc.
+     * bucket the test user has read+write access to.
+     */
+    public static void initialize(String bucketUrl, S3FileSystemOptions backendOptions) {
+        Objects.requireNonNull(bucketUrl, "bucketUrl");
+        Objects.requireNonNull(backendOptions, "backendOptions");
+
+        rootUrl = bucketUrl;
         options = backendOptions.toFileSystemOptions();
     }
 
@@ -54,6 +68,15 @@ public final class S3IntegrationContext {
     public static void reset() {
         rootUrl = null;
         options = null;
+    }
+
+    /**
+     * Random lowercase hex token suitable for embedding in a per-run S3 bucket name.
+     * Stable format (24 hex chars from 3 random ints) — matches what the URI-based
+     * {@code initialize} uses internally so all suites generate consistent names.
+     */
+    public static String randomBucketToken() {
+        return new Random().ints(3).mapToObj(i -> String.format("%08x", i)).collect(joining());
     }
 
     public static String rootUrl() {
