+++ modules/commons-vfs/src/integrationTest/java/com/github/vfss3/commonsvfs/local/CloudServerSuite.java
@@ -1,5 +1,7 @@
-package com.github.vfss3.commonsvfs;
+package com.github.vfss3.commonsvfs.local;
 
+import com.github.vfss3.commonsvfs.S3FileSystemOptions;
+import com.github.vfss3.commonsvfs.S3IntegrationContext;
 import java.net.URI;
 import org.junit.platform.suite.api.*;
 import org.testcontainers.containers.GenericContainer;
@@ -10,30 +12,28 @@ import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
 
 /**
  * Runs every test in {@code com.github.vfss3.commonsvfs.tests} against a freshly-started
- * MinIO container. Uses {@link GenericContainer} because the bundled testcontainers
- * version pre-dates the dedicated {@code MinIOContainer} module — once we bump
- * testcontainers we can swap to that.
+ * Zenko CloudServer container with the in-memory backend.
  */
 @Suite
-@SuiteDisplayName("MinIO integration tests")
+@SuiteDisplayName("CloudServer integration tests")
 @SelectPackages("com.github.vfss3.commonsvfs.tests")
-public class MinioSuite {
-    private static final DockerImageName MINIO_IMAGE =
-            DockerImageName.parse("minio/minio:RELEASE.2025-09-07T16-13-09Z");
-    private static final String ACCESS_KEY = "minioadmin";
-    private static final String SECRET_KEY = "minioadmin";
-    private static final int API_PORT = 9000;
+public class CloudServerSuite {
+    private static final DockerImageName IMAGE = DockerImageName.parse("zenko/cloudserver:latest-7.70.10");
+    private static final int API_PORT = 8000;
+    // Hardcoded in /conf/authdata.json baked into the image.
+    private static final String ACCESS_KEY = "accessKey1";
+    private static final String SECRET_KEY = "verySecretKey1";
 
     private static GenericContainer<?> container;
 
     @BeforeSuite
     static void startContainer() {
-        container = new GenericContainer<>(MINIO_IMAGE)
-                .withCommand("server", "/data")
-                .withEnv("MINIO_ROOT_USER", ACCESS_KEY)
-                .withEnv("MINIO_ROOT_PASSWORD", SECRET_KEY)
+        container = new GenericContainer<>(IMAGE)
+                .withEnv("S3BACKEND", "mem")
+                .withEnv("REMOTE_MANAGEMENT_DISABLE", "1")
+                .withEnv("ENDPOINT", "localhost")
                 .withExposedPorts(API_PORT)
-                .waitingFor(Wait.forHttp("/minio/health/live").forPort(API_PORT));
+                .waitingFor(Wait.forListeningPort());
         container.start();
 
         S3FileSystemOptions options = new S3FileSystemOptions();
