+++ modules/commons-vfs/src/integrationTest/java/com/github/vfss3/commonsvfs/local/SeaweedFsSuite.java
@@ -1,5 +1,7 @@
-package com.github.vfss3.commonsvfs;
+package com.github.vfss3.commonsvfs.local;
 
+import com.github.vfss3.commonsvfs.S3FileSystemOptions;
+import com.github.vfss3.commonsvfs.S3IntegrationContext;
 import java.net.URI;
 import org.junit.platform.suite.api.*;
 import org.testcontainers.containers.GenericContainer;
@@ -10,37 +12,32 @@ import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
 
 /**
  * Runs every test in {@code com.github.vfss3.commonsvfs.tests} against a freshly-started
- * Zenko CloudServer container with the in-memory backend.
+ * SeaweedFS container in S3 gateway mode.
  */
 @Suite
-@SuiteDisplayName("CloudServer integration tests")
+@SuiteDisplayName("SeaweedFS integration tests")
 @SelectPackages("com.github.vfss3.commonsvfs.tests")
-public class CloudServerSuite {
-    private static final DockerImageName IMAGE = DockerImageName.parse("zenko/cloudserver:latest-7.70.10");
-    private static final int API_PORT = 8000;
-    // Hardcoded in /conf/authdata.json baked into the image.
-    private static final String ACCESS_KEY = "accessKey1";
-    private static final String SECRET_KEY = "verySecretKey1";
+public class SeaweedFsSuite {
+    private static final DockerImageName IMAGE = DockerImageName.parse("chrislusf/seaweedfs:4.22");
+    private static final int S3_PORT = 8333;
 
     private static GenericContainer<?> container;
 
     @BeforeSuite
     static void startContainer() {
         container = new GenericContainer<>(IMAGE)
-                .withEnv("S3BACKEND", "mem")
-                .withEnv("REMOTE_MANAGEMENT_DISABLE", "1")
-                .withEnv("ENDPOINT", "localhost")
-                .withExposedPorts(API_PORT)
+                .withCommand("server", "-s3", "-dir=/data")
+                .withExposedPorts(S3_PORT)
                 .waitingFor(Wait.forListeningPort());
         container.start();
 
+        // SeaweedFS S3 API runs anonymous by default — any access key is accepted.
         S3FileSystemOptions options = new S3FileSystemOptions();
-        options.setCredentialsProvider(
-                StaticCredentialsProvider.create(AwsBasicCredentials.create(ACCESS_KEY, SECRET_KEY)));
+        options.setCredentialsProvider(StaticCredentialsProvider.create(AwsBasicCredentials.create("any", "any")));
         options.setUseHttps(false);
         options.setDisableChunkedEncoding(true);
 
-        URI endpoint = URI.create("http://" + container.getHost() + ":" + container.getMappedPort(API_PORT));
+        URI endpoint = URI.create("http://" + container.getHost() + ":" + container.getMappedPort(S3_PORT));
         S3IntegrationContext.initialize(endpoint, options);
     }
